pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ct-agent-frontdoor"
version = "0.1.0"
description = "CT image degradation-aware agent: perception → planning → multi-tool restoration → diagnosis → evaluation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-image>=0.20",
    "pillow>=9.5",
    "nibabel>=5.0",
    "pydicom>=2.4",
    "pyyaml>=6.0",
    "omegaconf>=2.3",
    "tqdm>=4.65",
]

[project.optional-dependencies]
iqa = [
    "pyiqa>=0.1.10",
]
dev = [
    "pytest>=7.3",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
